BOAC-5377: improves sidebar
