+++ src/assets/styles/global.css
@@ -328,15 +328,15 @@
 a {
   color: rgb(var(--v-theme-anchor));
   text-decoration: none;
-  background-color: transparent
+  background-color: transparent;
 }
 a:hover {
   color: #0056b3;
-  text-decoration: underline
+  text-decoration: underline;
 }
 body,
 html {
-  height: 100%
+  height: 100%;
 }
 h1 {
   font-size: 24px;
